splitting functions in multiple files and adding results processing in menu.
